% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erbs_decomposition.R
\name{clearness_index}
\alias{clearness_index}
\title{Clearness Index}
\usage{
clearness_index(
  ghi,
  solar_zenith,
  extra_radiation,
  min_cos_zenith = 0.065,
  max_clearness_index = 2
)
}
\arguments{
\item{ghi}{Global horizontal irradiance (W/m^2)}

\item{solar_zenith}{Solar zenith angle in degrees}

\item{extra_radiation}{Extraterrestrial radiation (W/m^2)}

\item{min_cos_zenith}{Minimum value of cos(zenith) to allow. Default: 0.065}

\item{max_clearness_index}{Maximum value of the clearness index. Default: 2.0}
}
\value{
Clearness index \link{unitless}
}
\description{
Calculate the clearness index, the ratio of global to extraterrestrial
irradiance on a horizontal plane.
}
\references{
Maxwell, E. L. (1987). A Quasi-Physical Model for Converting Hourly
Global Horizontal to Direct Normal Insolation. Technical Report No.
SERI/TR-215-3087, Solar Energy Research Institute.
}
\keyword{internal}
