% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erbs_decomposition.R
\name{erbs_decomposition}
\alias{erbs_decomposition}
\title{Erbs Decomposition of GHI into DNI and DHI}
\usage{
erbs_decomposition(
  time,
  lat,
  lon,
  GHI,
  min_cos_zenith = 0.065,
  max_zenith = 87,
  solar_constant = 1366.1
)
}
\arguments{
\item{time}{POSIXct vector of times (UTC recommended)}

\item{lat}{Latitude in degrees}

\item{lon}{Longitude in degrees}

\item{GHI}{Global horizontal irradiance (W/m^2)}

\item{min_cos_zenith}{Minimum value of cos(zenith) to allow when calculating
global clearness index kt. Equivalent to zenith = 86.273 degrees.
Default: 0.065}

\item{max_zenith}{Maximum value of zenith to allow in DNI calculation.
DNI will be set to 0 for times with zenith values greater than max_zenith.
Default: 87}

\item{solar_constant}{The solar constant (W/m^2). Default: 1366.1}
}
\value{
Data frame with columns:
\describe{
\item{time}{Input timestamps}
\item{GHI}{Input global horizontal irradiance (W/m^2)}
\item{DNI}{Direct normal irradiance (W/m^2)}
\item{DHI}{Diffuse horizontal irradiance (W/m^2)}
\item{kt}{Clearness index \link{unitless}}
\item{zenith}{Solar zenith angle (degrees)}
}
}
\description{
Estimate DNI and DHI from GHI using the Erbs model (Erbs et al., 1982).

The Erbs model estimates the diffuse fraction (DF) from global horizontal
irradiance through an empirical relationship between DF and the ratio of
GHI to extraterrestrial irradiance (Kt).

\deqn{DHI = DF \times GHI}
\deqn{DNI = (GHI - DHI) / \cos(\theta_z)}

where \eqn{\theta_z} is the solar zenith angle.
}
\examples{
\dontrun{
time <- seq(as.POSIXct("2026-01-15 06:00", tz = "UTC"),
            by = "hour", length.out = 12)
GHI <- c(50, 200, 450, 700, 850, 950, 1000, 950, 850, 700, 450, 200)

result <- erbs_decomposition(
  time = time,
  lat = -30.6279,
  lon = 24.0054,
  GHI = GHI
)
}

}
\references{
Erbs, D. G., Klein, S. A., & Duffie, J. A. (1982). Estimation of the
diffuse radiation fraction for hourly, daily and monthly-average global
radiation. Solar Energy, 28(4), 293-302.
\doi{10.1016/0038-092X(82)90302-4}
}
