% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erbs_decomposition.R
\name{get_extra_radiation_spencer}
\alias{get_extra_radiation_spencer}
\title{Extraterrestrial Radiation (Spencer Method)}
\usage{
get_extra_radiation_spencer(doy, solar_constant = 1366.1)
}
\arguments{
\item{doy}{Day of year (numeric or vector)}

\item{solar_constant}{The solar constant (W/m^2). Default: 1366.1}
}
\value{
Extraterrestrial radiation normal to the sun (W/m^2)
}
\description{
Determine extraterrestrial radiation from day of year using the
Spencer (1971) method.
}
\references{
Spencer, J. W. (1971). Fourier series representation of the sun. Search, 2, 172.
}
\keyword{internal}
