% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_linke_turbidity.R
\name{interpolate_turbidity}
\alias{interpolate_turbidity}
\title{Interpolate Monthly Turbidity to Daily Values}
\usage{
interpolate_turbidity(monthly_values, time_lt)
}
\arguments{
\item{monthly_values}{Numeric vector of length 12 with monthly turbidity values}

\item{time_lt}{POSIXlt object with timestamps}
}
\value{
Numeric vector of daily interpolated turbidity values
}
\description{
Interpolate 12 monthly Linke turbidity values to daily resolution using
day of year and sinusoidal fit through month midpoints.
}
\keyword{internal}
