% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_linke_turbidity.R
\name{linke_turbidity_filepath}
\alias{linke_turbidity_filepath}
\title{Get Path to LinkeTurbidities Database}
\usage{
linke_turbidity_filepath()
}
\value{
Character string with full path to LinkeTurbidities.h5 file
}
\description{
Returns the file path to the LinkeTurbidities.h5 database file
distributed with the package.
}
\examples{
\dontrun{
# Get path to database file
db_path <- linke_turbidity_filepath()
file.exists(db_path)  # Should be TRUE
}

}
