% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{restore_time_tz}
\alias{restore_time_tz}
\title{Restore Time to Original Timezone}
\usage{
restore_time_tz(time_utc, original_tz)
}
\arguments{
\item{time_utc}{POSIXct vector in UTC timezone.}

\item{original_tz}{Character string of the target timezone.}
}
\value{
POSIXct vector in the original timezone.
}
\description{
Converts UTC timestamps back to the original timezone for output.
}
\examples{
\dontrun{
# After calculations in UTC, restore to original SAST timezone
time_utc <- as.POSIXct("2026-01-15 10:00", tz = "UTC")
time_sast <- restore_time_tz(time_utc, "Africa/Johannesburg")
# time_sast shows 12:00 SAST
}

}
\keyword{internal}
